+++ R/core_x-to-monad.R
@@ -12,10 +12,9 @@
 #' \code{funnel} evaluates multiple arguments into one Rmonad. It can be used
 #' within pipelines to create multi-input nodes (works well with \code{\%*>\%}).
 #'
-#' \code{combine} takes a list and joins the elements into one Rmonad. If
-#' passed a list of non-monadic values, this will create a new monad that wraps
-#' a list, but it records each element of the list as a parent monad. It only
-#' really makes sense to use \code{combine} on lists of monads.
+#' \code{combine} takes a list of Rmonads and joins the elements into one
+#' Rmonad. The values of the original monadic containers joined into a list in
+#' the child Rmonad. The list Rmonads are recorded as the new Rmonad's parents.
 #'
 #' @param expr An expression
 #' @param xs  A list of elements to join into a monad
@@ -169,7 +168,9 @@ funnel <- function(..., env=parent.frame(), keep_history=TRUE){
 combine <- function(xs, keep_history=TRUE, desc=NULL){
 # combine :: [m *] -> m [*]
 
-  xs <- lapply(xs, as_monad)
+  if(!all(sapply(xs, is_rmonad))){
+    stop("'combine' works only on lists of Rmonad objects")
+  }
 
   # make a new monad that is the child of all monads in the input list
   out <- new_monad()